Added /register and /login routes. (POST requests)

+++ TheGame/TheGame/User.cpp
@@ -10,7 +10,7 @@ int User::GetId() const { return m_id; }
 const std::string& User::GetUsername() const { return m_username; }
 const std::string& User::GetPasswordHash() const { return m_password_hash; }
 int User::GetScore() const { return m_score; }
-int User::GetHoursPlayed() const { return m_hours_played; }
+double User::GetHoursPlayed() const { return m_hours_played; }
 
 void User::SetUsername(const std::string& username) { m_username = username; }
 void User::SetId(int id) { m_id = id; }
